+++ src/main/java/com/bettercontent/latentchemlib/block/ChemicalCloudBlock.java
@@ -46,7 +46,7 @@ public class ChemicalCloudBlock extends BaseEntityBlock {
 
     @Override
     public RenderShape getRenderShape(BlockState state) {
-        return RenderShape.INVISIBLE;
+        return RenderShape.MODEL;
     }
 
     @Override